Don't run hidden test if they only match exclusions
If a test is hidden using the [.] tag, it is surprising that it will run
if an exclusion matches, such as if the test executable is run with the
pattern '~[something]'.

Strengthen the semantics of [.] so that the test is only run if it
postively matches an included pattern.

This means that a test tagged as '[.foo]' will now match the pattern
'[foo]'. but not the pattern '~[bar]'